Handle empty iobuf in Thrift checksum adapter
Summary:
During S577904 investigation we suspect that crash is happening because nonPayload info is corrupted as ioBuf for nonPayload is set to null. We do not check if receiving ioBuf or nonPayload's iobuf are valid before accessing them.
This diff doesn't fix the sev issue but we should not crash client process rather graceful handle empty iobuf by throwing exceptions and logging.

Differential Revision: D85472448

fbshipit-source-id: 623d3c68620b6625d00b29254784db3895c8167f