fix BadEnum bug in converter
Summary:
As part of migration, customers are using `_to_py_deprecated()` to convert already migrated thrift-python values back to py-deprecated in unmigrated code.

However, the converter isn't converting `BadEnum`. Rather than leaving it as a thrift-python concept that will serialize weirdly, we should convert it to a py-deprecated enum, aka an `int`.

Reviewed By: romaad, yoney

Differential Revision: D85723505

fbshipit-source-id: 70ca21e55d44cf27e35a40c8e8f9b739643414cc