thrift-py3: make tests python 3.14 ready
Summary:
asyncio.get_event_loop() outside of a `async def` is now very suspect since there is no such thing as a default eventloop.   Convert all theses tests to IsolatedAsyncioTestCase

also deleted that `test_queue_timeout` test after trying to fix it. it was broken in 3.14 and already skipped in 3.12 and 3.10

Reviewed By: ahilger

Differential Revision: D85872160

fbshipit-source-id: 72455a3b90810e5472cd30a1dc06cdfe6fb44855