Fix propagation of load header from HTTP server
Summary:
_sustained internal screaming_

HTTP clients expect the `load` header (eg for load polling in SR), so this diff adds it. This wasn't an issue with the rocket transport, because we send the metadata inline in the request (instead of needing to translate it back/forth from headers).

These seem to be the only things we copy from header in

https://www.internalfb.com/code/fbsource/[447b9e452caf0ce22243d951157eef2129e12c54]/fbcode/thrift/lib/cpp2/transport/core/RpcMetadataUtil.cpp?lines=194-195

 and

https://www.internalfb.com/code/fbsource/[447b9e452caf0ce22243d951157eef2129e12c54]/fbcode/thrift/lib/cpp2/transport/core/RpcMetadataUtil.cpp?lines=248-249

So I think they are the only things we need to account for here.

Reviewed By: 3ach

Differential Revision: D85897283

fbshipit-source-id: 402dcf752402c6974748cbd5f2725ef4e29364da