fix undefined _fbthrift__py3_module_name bug
Summary:
This fixes `NameError: name '_fbthrift__py3_module_name' is not defined`.

The current template only defines this symbol when the IDL defines at least one type (`struct`, `union`, `typedef`, or `enum`). However, it's possible for an IDL to define no types but include `service` methods with containers. The symbol is needed if there are any containers, even containers of primitives.

So we move the variable name definition out of the `#program:has_types?` gate.

Reviewed By: stanleyz510

Differential Revision: D85972493

fbshipit-source-id: 2cf81e37159878fe2fbe07b7bd49a9d3f2209365