Fix client side connection close
Summary:
We don't do any cleanup for Bidi if the connection is closed while the stream is alive. Thid diff adds the cleanup - we propagate the stream error to the stream task and cancel the sink.

In practice cancelling the sink ends up being a noop because no messages are going to be sent - however we need to do this to clean up the ClientSinkBridge queue.

Reviewed By: tlj77, sazonovkirill

Differential Revision: D86121596

fbshipit-source-id: 844cd0ec771ae051c08164092ae3e796583a0963