Fix missing enum values in switch statements (xplat/thrift)
Summary:
We are rolling out `-Wswitch-enum` and `-Wswitch-default` in `fbobjc` to have all switches be exhausive by default.

This means either:

  1. Adding any missing enum values to the switch
  2. Opting the target or hierarchy of targets out of being exhaustive for now (`-Wno-switch-enum` & `-Wno-switch-default`)
  3. Opting a switch out of being exhaustive:
    * `NON_EXHAUSTIVE_SWITCH_BEGIN|END`
    * `FB_NON_EXHAUSTIVE_SWITCH_BEGIN|END`
    * `FOLLY_NON_EXHAUSTIVE_SWITCH({ ... })`
    * `C10_DIAGNOSTIC_PUSH_AND_IGNORED_IF_DEFINED("-Wswitch-enum")`
    * `#pragma GCC|clang diagnostic ignored "-Wswitch-enum"

`-Wswitch-enum` will make is so that when we add a new value to an enum, all relevant switch statements raise a compiler warning to remind us to update them (regardless of having a `default` case or not).

`-Wswitch-default` will enforce that we always have a default so that we can cover values that are outside the defined enum values.  Enums are just integers so switching on a value that is outside an enum definition is completely valid, but if nothing handles it we can get unexpected or even undefined behavior.  Better to `abort` than have impossible to debug undefined behavior.

See https://fburl.com/switch-enum for more.

Differential Revision: D86026847

fbshipit-source-id: beb4e59157e930138867afa5877ff638fe55c1d9