test helper conditional on float32 rollout
Summary:
This is an alternative to the annotation `python.EnableUnsafeUnconstrainedFloat32`. While the annotation can be used to fix multiple test cases at once, it may be overkill for cases where only 1-2 test cases need to be fixed.

Adding a rollout-conditional version of the `round_thrift_float32_if_rollout` allows pre-placing float32 rounding before rollout. It also bypasses the issue of thrift files that must be updated in configerator, because it may take a while for the configerator thrift compiler to ingest the annotation.

Reviewed By: prakashgayasen

Differential Revision: D86255909

fbshipit-source-id: 0a466ea3dad37ee0cb0c328e395aa7bdb74127bb