strongly type List and Set `_fbthrift_elements`
Summary:
We have the opportunity to strongly type the container `_fbthrift_elements` member if we move its declaration from `Container` base class to `List`, `Set`, and `Map` classes themselves.

This results in faster access in some situations; for example in `List.__getitems__`, we now use `PyTuple_GetItem` instead of `PyObject_GetItem`.

The perf improvements is expected to be small if measurable. But the code is now easier to maintain.

Reviewed By: Sunnie912

Differential Revision: D86342961

fbshipit-source-id: ce0b172ab6533d6a98bf9fe12733c83908a75189