fix trace size propagation for WWW, add depth
Summary:
# Context

Today, WWW does **not** participate in incrementing depth, like C++ services do via our TraceContinuationHandler. This means if we have a call path:

C++ (depth = 1) -> thrift/HHVM (depth = 1, BAD!) -> C++ (depth = 2)

the depth is wrong.

# This diff

Start ensuring HHVM is participating and incrementing depth.

# some things to address

- What if depth >= 100 in HHVM; should we not trace limit?

I doubt this is happening in practice but we can embed this functionality to trace continuation later if it is relevant.

-  What about breadthDepthProduct?

we'll need to eventually also care about BDP, but depth is P0 since our current C++ instrumentation trace limits on it

Differential Revision: D85785995

fbshipit-source-id: 7fefdb964af12aedf772d65f29e65f18f393454e