Make TypeRef trivially copyable
Summary:
The current design stores a boxed typeref to the element type of each container, which has to be cloned to copy the typeref. This diff instead stores the container type in a thread-safe cache on each instance of TypeSystem and has the TypeRef store a pointer to that, making TypeRef trivially copyable.
Some types are reordered to avoid needing an extra layer of boxing.

Reviewed By: praihan

Differential Revision: D86462591

fbshipit-source-id: 50e4a41d72f0c76d9a349fc35feda342c96a74fb