Refactor underlying implementation of `PythonAsyncProcessor::dispatchRequest*` to coroutines.
Summary:
As titled.

The motivation for this change is to make the follow-up diff easier to write. Specifically, the serializedRequest parameter would need to be used in multiple lambdas with the current SemiFuture implementation, and guaranteeing the lifetime across those lambdas turned out to be tricky (I got it wrong multiple times).

Instead of doing that the hard way, the whole implementation of the function can be converted to a coroutine inside a single lambda. With that, lifetime management of `serializedRequest` becomes trivial.

This also made it easy to collapse the duplicate calls to `handlePythonServerCallback*` into a single call.

ref: https://www.internalfb.com/wiki/Scuba_Internal_Team/Performance_and_Optimization/Future_to_Coroutine_Recipe_Book/

Reviewed By: ahilger

Differential Revision: D86481020

fbshipit-source-id: 32e96951ec87df2f9a4729c3cb45806dfc550653