Support capping the number of experiments we add to global state
Summary:
_something_ about ABS & direct thrift is sus here (although it is not
necessarily the only offender)

https://pxl.cl/8s4rX

The drop here is the canary + land of D86708621, although this is a pretty big
hammer. So, try to write something a bit more targeted, so we can move
debugging into PHP where it will be easier to reason about.

The behavior here has a JK (where a value of 0 is disabled), and if the current
state is over the limit, we do not add any more.

This is somewhat analgalous to the logic in C++ that trims the number of IDs
that we _send_ on outbound requests

https://www.internalfb.com/code/fbsource/[8e541b0c6e663cc82f6cda579a29afb5bbbfc9d3]/fbcode/common/thrift/cpp/server/thrift_framework_metadata/ContextPropComponent.cpp?lines=300-321

However, I can't figure out how the direct thrift server would impact this, but
there are some real CPU impacts on this list growing large
(https://fb.workplace.com/groups/wfi/posts/29966230699665493)

So, let's do this as the next intermediate step while we continue debugging.

Reviewed By: armanoz

Differential Revision: D86783900

fbshipit-source-id: 488bb9bc06f2f69e1e1b0edb8a71cca17ba386e7