move _fbthrift_data out of StructOrUnion
Summary:
This is first step to switch the `Struct` internal data from `tuple` to an object that encapsulates `tuple`. This will allow overriding `__eq__`, `__lt__`, and `__hash__` for faster comparisons.

In this diff, we move `_fbthrift_data` (aka internal data) from `StructOrUnion` to be `Struct` and `Union`. That allows us to encapsulate the internal data in `Struct` while leaving it as `tuple` for `Union`.

Reviewed By: createdbysk

Differential Revision: D87488378

fbshipit-source-id: 89dc091f1f1ddedc14fdbc85143b40b8e8db933b