Convert JSON structured annotation template to Whisker partial
Summary:
NOTE: Fixture changes are whitespace (indentation) only.

This is an artifact of how partials' preceding indentation levels are computed by the Whisker engine. Ideally the changed lines would be indented one more level, but the indent level of `"type":` is not taken because the partial line is not *standalone*.

Instead, it takes the parent indentation level from the `m_partials.structured_annotations` usages, which are standalone.

Reviewed By: iahs

Differential Revision: D87875034

fbshipit-source-id: a798147b8948c7fc782430754340295c0fcd5f80