only generate one version of types reflection
Summary:
All thrift-python libraries depend on -py3-types-reflection if `"py3"` is also in their `languages`.

However, we're generating two versions of py3-types reflection via auto-migrate constraint, even though the content of py3-types-reflection is the same in both cases.

To solve this, make a `-python-types` library unconditionally depend on `-py3-types-reflection` if `"py3"` is in `languages`. This should be ok from a build size perspective because `-py3-types-reflection` is a `python_library`, and it doesn't have any __generated__ `cython_library` dependencies.

Reviewed By: prakashgayasen

Differential Revision: D87494247

fbshipit-source-id: f4b851184f4d1782e2a19cf0e766dc4906e8be26