fix metadata bug for exception field in struct
Summary:
This fixes a bug triggered by iterating `fields` of `ThriftStructProxy`, when one of the fields is an exception.

The issue is that we don't have enough info in `ThriftStructType` to know before hand whether it's a struct or exception. The only way is checking whether the name is in the `structs` or `exceptions` maps at the top-level metadata object.

Since it's unusual to nest `exception` field in a `struct`, let's avoid the performance hit of pre-checking the map, and instead use `try-except KeyError` to handle the case where it's actually an exception.

Reviewed By: vitaut, prakashgayasen, aristidisp

Differential Revision: D88286562

fbshipit-source-id: 4fc02031050ea547013385a3805b1c8954e74a8d