Don't CHECK serverCallbackReady
Summary:
A streaming request can crash on the client if the server closes the connection before the first response is sent (repro D88216572).

From a quick read of the I would argue that `serverCallback == 0` can be safely interpreted as the stream never started. This should only be true between client callback construction and reception of the first response or cancellation. The tricky part would be that RocketStreamClientCallback::onFirstResponse accepts StreamServerCallback as a raw pointer, which implies that you could pass nullptr. I have a hard time believing that this code could possibly be correct if you passed `nullptr`.

Reviewed By: sazonovkirill

Differential Revision: D88283987

fbshipit-source-id: a5db18476c60637b6236ee111ea73a26e5da6d98