inline import in metadata.thrift
Summary:
It's possible to trigger import-order dependent `ImportError` due to the import of `apache.thrift.metadata.thrift_types` in `apache.thrift.metadata.thrift_enums`. This creates a circular reference because the `thrift_types` itself imports the `.thrift_enums`. For a repro of this, see D88907726.

This only affects `thrift/lib/thrift/metadata.thrift`.

We can remove this circularity by doing the import inline. This breaks the typing of the functions that return `_fbthrift_metadata.ThriftMetadata` type, so let's only apply the change to `thrift/lib/thrift/metadata.thrift` via compiler option, that will only be applied to `metadata.thrift`.

Reviewed By: Sunnie912, prakashgayasen

Differential Revision: D89006632

fbshipit-source-id: 532fab33803dd640a6d77c451ff7c6c4787c5b73