Add enum value to name converter with thrift-py-deprecated compatible error
Summary:
`ttypes.EnumName._VALUES_TO_NAMES[v]` throws a `KeyError` if `v` is not in `ttypes.EnumName._VALUES_TO_NAMES`.

`thrift_enums.EnumName(v).name` throws a `ValueError` in the same situation.

To make the codemod from the former to latter maintain the original `thrift-py-deprecated` (i.e. `ttypes`) behavior, create a function to raise a `KeyError` whenever the value is not part of the named values in the Enum.

Reviewed By: prakashgayasen

Differential Revision: D89110930

fbshipit-source-id: a0dd2faba1b4eaff459e150dc2a539b1988e0e84